separate interaction files, implement evm-chains, remove nickname
