+++ package.json
@@ -52,6 +52,7 @@
         "discord.js": "^13.6.0",
         "dotenv": "^10.0.0",
         "ethers": "^5.5.3",
+        "evm-chains": "^0.2.0",
         "form-data": "^4.0.0",
         "lodash.clonedeep": "^4.5.0",
         "lodash.isequal": "^4.5.0",
