+++ src/app/types/discord/ConnectedAddress.ts
@@ -1,5 +1,4 @@
 export interface ConnectedAddress {
     address: string,
     chainId: string,
-    nickname: string | null,
 }
