+++ src/app/types/interactions/DEGENMenuInteraction.ts
