+++ src/app/types/interactions/DiscordButtons.ts
@@ -1,0 +1,9 @@
+import { MessageButtonStyle } from 'discord.js';
+
+export interface DiscordButtons {
+    label: string,
+    style: MessageButtonStyle,
+    function: string,
+    successMessage: string,
+    failureMessage: string,
+}
