+++ src/app/utils/interactions/MenuInteractions/deleteUserAddressInteraction.ts
@@ -1,17 +1,18 @@
-import { sendMenuInteraction } from '../interactionBuilders/sendMenuInteraction';
+import { sendMenuInteraction } from '../../interactionBuilders/sendMenuInteraction';
 import { DMChannel, User } from 'discord.js';
-import { DiscordUserCollection } from '../../types/discord/DiscordUserCollection';
-import { ConnectedAddress } from '../../types/discord/ConnectedAddress';
+import { DiscordUserCollection } from '../../../types/discord/DiscordUserCollection';
+import { ConnectedAddress } from '../../../types/discord/ConnectedAddress';
 import { MessageSelectOptionData } from 'discord.js';
-import { DEGENMenuInteraction } from '../../types/DEGENMenuInteraction';
-import Log from '../Log';
+import { DEGENMenuInteraction } from '../../../types/interactions/DEGENMenuInteraction';
+import Log from '../../Log';
+import { getChain } from 'evm-chains';
 
 export const deletUserAddressInteraction = (user: User, dmChannel:DMChannel, discordUserDocument: DiscordUserCollection, connectedAddresses: ConnectedAddress[]): Promise<DEGENMenuInteraction> => {
 	Log.debug('deleteUserAddressInteraction started.');
 
 	const menuItems: MessageSelectOptionData[] = connectedAddresses.map((address, i) => {
 		return {
-			label: `${address.nickname}`,
+			label: `${getChain(parseInt(address.chainId)).name}`,
 			description: `ChainId: ${address.chainId}, Address: ${address.address}`,
 			value: i.toString(),
 		};
