+++ src/app/utils/interactions/functionTable.ts
@@ -1,30 +1,46 @@
 import { updatePOAPDAddress } from './../mongoDbOperations/updatePOAPDeliveryAddress';
-import { deletUserAddressInteraction } from './../interactions/deleteUserAddressInteraction';
+import { deletUserAddressInteraction } from './MenuInteractions/deleteUserAddressInteraction';
 import { deleteUserAddress } from './../mongoDbOperations/deleteUserAddress';
 import { v1WalletConnect } from '../v1WalletConnect';
 import { sendLinkToWebsite } from '../sendLinkToWebsite';
-import { changePOAPAddressInteraction } from '../interactions/changePOAPAddressInteraction';
-import Log from '../Log';
+import { changePOAPAddressInteraction } from './MenuInteractions/changePOAPAddressInteraction';
+import Log, { LogUtils } from '../Log';
+import { DMChannel, User } from 'discord.js';
+import { DiscordUserCollection } from '../../types/discord/DiscordUserCollection';
+import { error } from 'console';
 
-export const functionTable = async (functionToCall: string, args:any, selectedItemDescription?:string): Promise<any> => {
+export const functionTable = async (functionToCall: string, args:{user: User, dmChannel: DMChannel, discordUserDocument: DiscordUserCollection}, selectedItemDescription?:string): Promise<any> => {
 	Log.debug(`FunctionTable ${functionToCall}`);
 	const { user, dmChannel, discordUserDocument } = args;
 
-	selectedItemDescription === undefined ? '' : selectedItemDescription;
+	selectedItemDescription === undefined ? 'empty' : selectedItemDescription;
 
 	const connectedAddresses = args.discordUserDocument?.connectedAddresses ? args.discordUserDocument.connectedAddresses : null;
-	if (functionToCall === 'walletConnect') {
+	
+	switch (functionToCall) {
+	case 'walletConnect':
 		// const nickname = await chooseANickName(user, dmChannel, connectedAddresses);
 		return await v1WalletConnect(user, dmChannel, connectedAddresses);
-	} if (functionToCall === 'changePOAPAddress' && discordUserDocument && connectedAddresses) {
-		return await changePOAPAddressInteraction(user, dmChannel, discordUserDocument, connectedAddresses);
-	} if (functionToCall === 'updatePOAPAddress' && connectedAddresses && selectedItemDescription) {
-		return await updatePOAPDAddress(user, selectedItemDescription);
-	} if (functionToCall === 'deleteAddressInteraction' && connectedAddresses && discordUserDocument) {
-		return await deletUserAddressInteraction(user, dmChannel, discordUserDocument, connectedAddresses);
-	} if (functionToCall === 'deleteUserAddress' && selectedItemDescription) {
-		return await deleteUserAddress(user, selectedItemDescription);
-	} else {
+	case 'changePOAPAddress':
+		if (connectedAddresses) {
+			return await changePOAPAddressInteraction(user, dmChannel, discordUserDocument, connectedAddresses);
+		} break;
+	case 'deleteAddressInteraction':
+		if (connectedAddresses) {
+			return await deletUserAddressInteraction(user, dmChannel, discordUserDocument, connectedAddresses);
+		} break;
+	case 'updatePOAPAddress':
+		if (selectedItemDescription) {
+			return await updatePOAPDAddress(user, selectedItemDescription);
+		} break;
+	case 'deleteUserAddress':
+		if (selectedItemDescription) {
+			return await deleteUserAddress(user, selectedItemDescription);
+		} break;
+	case 'sendLinkToWebsite':
 		return await sendLinkToWebsite();
+	default:
+		LogUtils.logError('Function table did not find a case.', error);
+		await dmChannel.send('Something appears to have gone wrong.');
 	}
 };
